+++ src/mock/mock.json
@@ -9,6 +9,7 @@
       "published": "29 May 2020",
       "category": "computers",
       "ownerId": "test-user",
+      "rating": 3.4,
       "votes": {
         "jack": 3,
         "jane": 3,
@@ -45,6 +46,7 @@
       "published": "4 Feb. 2008",
       "category": "computers",
       "ownerId": "test-user",
+      "rating": 3.6,
       "votes": {
         "jack": 3,
         "max": 2,
@@ -78,6 +80,7 @@
       "published": "18 May 2018",
       "category": "travel",
       "ownerId": "test-user",
+      "rating": 0,
       "votes": {},
       "comments": [
         {
@@ -97,6 +100,7 @@
       "published": "23 Dec. 2010",
       "category": "computers",
       "ownerId": "test-user",
+      "rating": 0,
       "votes": {},
       "comments": [
         {
@@ -116,6 +120,7 @@
       "published": "27 Oct. 2012",
       "category": "art",
       "ownerId": "jack",
+      "rating": 4.1,
       "votes": {
         "jack": 5,
         "donald": 5,
@@ -150,6 +155,7 @@
       "published": "6 Aug. 2016",
       "category": "art",
       "ownerId": "jack",
+      "rating": 3.7,
       "votes": {
         "jack": 3,
         "max": 3,
@@ -173,6 +179,7 @@
       "published": "13 April 2009",
       "category": "art",
       "ownerId": "jack",
+      "rating": 4,
       "votes": {
         "jack": 4,
         "max": 3,
@@ -202,6 +209,7 @@
       "published": "10 Nov. 2017",
       "category": "travel",
       "ownerId": "ann",
+      "rating": 2.5,
       "votes": {"elodie": 2, "ann": 3 },
       "comments": [
         {
@@ -227,7 +235,7 @@
       "published": "27 Aug. 2020",
       "category": "travel",
       "ownerId": "ann",
-
+      "rating": 3.1,
       "votes": {
         "jack": 5,
         "franz": 3,
@@ -249,6 +257,7 @@
       "published": "11 Sept. 2020",
       "category": "travel",
       "ownerId": "ann",
+      "rating": 2,
       "votes": { "elodie": 2 },
       "comments": [
         {
@@ -268,6 +277,7 @@
       "published": "29 Oct. 2019",
       "category": "computers",
       "ownerId": "test-user",
+      "rating": 2,
       "votes": { "test-user": 2 },
       "comments": [
         {
@@ -293,6 +303,7 @@
       "published": "18 Dec. 2020",
       "category": "computers",
       "ownerId": "test-user",
+      "rating": 3,
       "votes": { "test-user": 3 },
       "comments": [
         {
@@ -318,6 +329,7 @@
       "published": "23 Feb. 2012",
       "category": "education",
       "ownerId": "mike",
+      "rating": 3,
       "votes": { "mike": 3 },
       "comments": []
     },
@@ -330,6 +342,7 @@
       "published": "21 Jun. 2018",
       "category": "education",
       "ownerId": "mike",
+      "rating": 0,
       "votes": {},
       "comments": []
     }
